git-svn-id: svn://192.168.219.100/torrent-auto-downloader@925 eb2df1e2-efe3-6949-8084-3fc849c3e4eb

+++ trunk/solution_private/TorrentAutoDownloader/src/TorrentAD/src/main/java/kr/co/darkkaiser/torrentad/net/torrent/transmission/methodresult/TorrentGetMethodResult.java
@@ -13,17 +13,17 @@ public final class TorrentGetMethodResult extends AbstractMethodResult {
 	public final class Torrent {
 
 		private long id;
-		private long status;
+		private int status;		// 0:Stopped, 1:Check waiting, 2:Checking, 3:Download waiting, 4:Downloading, 5:Seed waiting, 6:Seeding
 		private boolean isStalled;
 		private boolean isFinished;
-		private long error;
+		private int error;
 		private String errorString;
 
 		public long getId() {
 			return this.id;
 		}
 		
-		public long status() {
+		public int status() {
 			return this.status;
 		}
 		
@@ -35,7 +35,7 @@ public final class TorrentGetMethodResult extends AbstractMethodResult {
 			return this.isFinished;
 		}
 		
-		public long error() {
+		public int error() {
 			return this.error;
 		}
